Change written format, Implement basic Place and Move rules
The format makes it easier to implement, and I think may also make more
sense. It definitely makes the common moves shorter to write.

So far I have the rules that you can't place a piece onto another piece,
and that you can only move (and have to "move" all of - even for 0
distance) the pieces on the place you move from.